[mob] refactor
