+++ mobile/lib/services/file_magic_service.dart
@@ -17,7 +17,7 @@ import "package:photos/models/metadata/common_keys.dart";
 import "package:photos/models/metadata/file_magic.dart";
 import 'package:photos/services/remote_sync_service.dart';
 import 'package:photos/utils/crypto_util.dart';
-import 'package:photos/utils/file_download_util.dart';
+import "package:photos/utils/file_key.dart";
 
 class FileMagicService {
   final _logger = Logger("FileMagicService");
