+++ mobile/lib/utils/file_uploader.dart
@@ -38,7 +38,7 @@ import 'package:photos/services/local_sync_service.dart';
 import 'package:photos/services/sync_service.dart';
 import "package:photos/services/user_service.dart";
 import 'package:photos/utils/crypto_util.dart';
-import 'package:photos/utils/file_download_util.dart';
+import "package:photos/utils/file_key.dart";
 import 'package:photos/utils/file_uploader_util.dart';
 import "package:photos/utils/file_util.dart";
 import "package:photos/utils/network_util.dart";
